refactor!(Client): Moved back to a single file.
